adress requested changes

+++ margin_app/app/schemas/order.py
@@ -3,6 +3,7 @@ Pydantic schemas for order operations.
 """
 
 import uuid
+from decimal import Decimal
 
 from pydantic import BaseModel, Field
 
@@ -11,7 +12,7 @@ class OrderCreate(BaseModel):
     """Schema for creating a new order"""
 
     user_id: uuid.UUID = Field(..., description="ID of the user placing the order")
-    price: float = Field(..., description="Price of the order")
+    price: Decimal = Field(..., description="Price of the order")
     token: str = Field(..., description="Token symbol for the order")
     position: uuid.UUID = Field(..., description="Position ID related to the order")
 
@@ -21,7 +22,7 @@ class OrderResponse(BaseModel):
 
     id: uuid.UUID
     user_id: uuid.UUID
-    price: float
+    price: Decimal
     token: str
     position: uuid.UUID
